feat: frontend 디렉토리 추가, 페이지 경로 추가 [HOME-ALONEn]
