+++ p4th/HeejiYun/frontend/lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/pages/home_page.dart';
+import 'package:frontend/pages/sign_in_page.dart';
+import 'package:frontend/pages/sign_up_complete_page.dart';
 import 'package:frontend/pages/sign_up_page.dart';
 
 void main() {
@@ -17,16 +19,18 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
           textButtonTheme: TextButtonThemeData(
               style: TextButton.styleFrom(
-        backgroundColor: Colors.cyanAccent.shade100,
+        backgroundColor: Colors.cyanAccent.shade200,
         primary: Colors.white,
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
         minimumSize: Size(300, 45),
       ))),
       title: 'HOME ALONE',
-      initialRoute: "/sign-up",
+      initialRoute: "/sign-up-complete",
       routes: {
         "/sign-up": (context) => SignUpPage(),
         "/home": (context) => HomePage(),
+        "/sign-in": (context) => SignInPage(),
+        "/sign-up-complete": (context) =>SignUpCompletePage(),
       },
     );
   }
